DBC22-2376: header rework
